+++ liquid/utils.py
@@ -165,15 +165,33 @@ def _exc_stack_code(context):
     console = Console(file=StringIO())
     console.print(f"{context.path!r}, line {context.lineno + 1}, "
                   f"column {context.colno + 1}")
-    if not hasattr(context.stream, 'seekable') or not context.stream.seekable():
+
+    stream = context.stream
+    seekable = False
+    if stream.closed:
+        try:
+            if isinstance(stream, StringIO):
+                stream = StringIO(stream.getvalue())
+            else:
+                stream = open(stream.name)
+            seekable = stream.seekable()
+        except (AttributeError, FileNotFoundError, IOError, ValueError):
+            seekable = False
+    else:
+        try:
+            seekable = stream.seekable()
+        except (AttributeError, ValueError, IOError):
+            seekable = False
+
+    if not seekable:
         console.print("  [Stream not seekable]") # pragma: no cover
     else:
-        context.stream.seek(0)
+        stream.seek(0)
         line_range = (
             max(0, context.lineno - LIQUID_EXC_CODE_CONTEXT) + 1,
             context.lineno + LIQUID_EXC_CODE_CONTEXT + 1
         )
-        code = Syntax(context.stream.read(),
+        code = Syntax(stream.read(),
                       lexer_name='liquid',
                       line_numbers=True,
                       line_range=line_range,
@@ -193,7 +211,7 @@ def excmsg_with_context(msg, context, parser):
         The assembled exception message
     """
     config = parser.config if parser else None
-    if not context or not config or not config.debug:
+    if not context or not config: # or not config.debug:
         return msg
 
     stacks = [context]