+++ tests/test_python/test_tags.py
@@ -230,6 +230,17 @@ def test_for():
         {%- endfor -%}
     ''', {'mode': 'python'}).render() == '0'
 
+    assert Liquid('''{% for i in 3 | range -%}
+        {{i}}{% break %}
+        {%- else -%}x
+        {%- endfor -%}
+    ''', {'mode': 'python'}).render() == '0'
+
+    assert Liquid('''{% for key, val in d.items() -%}
+        {{key}}-{{val}}{% break %}
+        {%- endfor -%}
+    ''', {'mode': 'python'}, d={'a': 1}).render().strip() == 'a-1'
+
 def test_while():
 
     assert Liquid('''